Support Process.Executable{Path,Args} on macOS ≤ 10.15
On macOS before 10.15, sysctl("kern.procargs2") doesn't seem to return
the process executable and arguments in the same format as on later
macOS version. Instead, on macOS ≤ 10.15 just use the number of
arguments (which is correct) and then fetch the actual arguments using
sysctl("kern.procargs") and use argc to detect the argv/envp boundary.

Idea inspired by the implementation in mesos, see
http://git-wip-us.apache.org/repos/asf/mesos/commit/97ed2e5d